convert the color picker button input to CDS buttonmain and move to better file location [CV2-3360]
