+++ src/app/components/media/ReportDesigner/ReportDesigner.module.css
@@ -98,9 +98,7 @@
               gap: 8px;
 
               > button {
-                flex: 0 0 48px;
-                height: 48px;
-                margin: 23px 0 0;
+                align-self: center;
               }
 
               > div {
