+++ src/app/components/media/ReportDesigner/ReportDesignerForm.js
@@ -5,7 +5,7 @@ import cx from 'classnames/bind';
 import ButtonMain from '../../cds/buttons-checkboxes-chips/ButtonMain';
 import TextArea from '../../cds/inputs/TextArea';
 import { ToggleButton, ToggleButtonGroup } from '../../cds/inputs/ToggleButtonGroup';
-import ColorPicker from '../../layout/ColorPicker';
+import ColorPicker from '../../cds/inputs/ColorPicker';
 import UploadFile from '../../UploadFile';
 import SwitchComponent from '../../cds/inputs/SwitchComponent';
 import LanguagePickerSelect from '../../cds/inputs/LanguagePickerSelect';
